feat: rename network utils
